+++ CHANGELOG.md
@@ -53,6 +53,8 @@ For instructions on upgrading to newer versions, visit
 
 * \#3868 Loading models in rake tasks now expands the rails path.
 
+* \#3764 Fixed case statement check for enumerable targets.
+
 * \#3740 Fixes `Missing attribute: '_id'` error when using methods only or without (dx7)
 
 * \#3631 Fixes issue where `before_save` callback can get called twice after a child create